Refactor: Implement global error handlers.

+++ packages/backend/services/ExpenseService.ts
@@ -1,5 +1,6 @@
 import type { Expense } from "@prisma/client";
-import { PrismaService } from "./PrismaService";
+import { PrismaService } from "./";
+import { BadRequestError, NotFoundError } from "../errors";
 
 export interface ExpenseDto {
   title: string;
@@ -22,14 +23,18 @@ export class ExpenseService {
     });
   }
 
-  public async getById(id: number): Promise<Expense | null> {
-    return this.prisma.expense.findUnique({
+  public async getById(id: number): Promise<Expense> {
+    const expense = await this.prisma.expense.findUnique({
       where: { id: id },
     });
+    if (!expense) {
+      throw new NotFoundError("Expense not found");
+    }
+    return expense;
   }
 
   public async create(data: ExpenseDto): Promise<Expense> {
-    return this.prisma.expense.create({
+    const expense = await this.prisma.expense.create({
       data: {
         title: data.title,
         note: data.note ?? "",
@@ -39,19 +44,31 @@ export class ExpenseService {
         userId: data.userId,
       },
     });
+    if (!expense) {
+      throw new BadRequestError();
+    }
+    return expense;
   }
 
   public async update(
     id: number,
     data: Partial<Pick<Expense, "title" | "note" | "amount" | "category">>,
   ): Promise<Expense> {
-    return this.prisma.expense.update({
+    const expense = await this.prisma.expense.update({
       where: { id: id },
       data,
     });
+    if (!expense) {
+      throw new NotFoundError("Expense not found");
+    }
+    return expense;
   }
 
   public async delete(id: number): Promise<void> {
+    const expense = await this.getById(id);
+    if (!expense) {
+      throw new NotFoundError("Expense not found");
+    }
     await this.prisma.expense.delete({ where: { id: id } });
   }
 }
